LayoutTests/fast/css/image-value-type.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../js/resources/js-test-pre.js"></script>
<style>
div#test {
    background-image: url(stars.gif);
}
</style>
</head>
<body>
<script>

description("Documents that CSS image values are not primitive values anymore.");

var style = document.styleSheets[1].rules[0].style;
var imageValue = style.getPropertyCSSValue('background-image');

shouldBe("imageValue.cssValueType", "CSSValue.CSS_PRIMITIVE_VALUE");

</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
